+++ test/squint/compiler_test.cljs
@@ -1851,6 +1851,10 @@
   (is (false? (jsv! "(not \"\")")))
   (is (not (str/includes? (jss! "(not (zero? 1))") "not"))))
 
+(deftest reduce-kv-test
+  (is (eq {1 :a, 2 :b, 3 :c} (jsv! "(reduce-kv #(assoc %1 %3 %2) {} {:a 1 :b 2 :c 3})")))
+  (is (eq {1 :a, 2 :b, 3 :c} (jsv! "(reduce-kv #(assoc %1 %3 %2) {} (new js/Map (js/Object.entries {:a 1 :b 2 :c 3})))"))))
+
 (defn wrap-async [s]
   (str/replace "(async function () {\n%s\n})()" "%s" s))