+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Unreleased
 
 - Add `clojure.set` functions `select`, and `rename-keys`
+- Fix `reduce-kv` with `js/Map` input
 
 ## 0.4.76 (2023-12-07)